Fix: 수정된 타입을 api 및 customhook Mutation에 적용

+++ pages/api/auth/deleteToken/index.ts
@@ -1,6 +1,7 @@
 import { deleteToken } from "@/lib/api/redis";
-import { DBClient } from "@/lib/database";
-import { UserData } from "@/types/apiTypes";
+import dbConnect from "@/lib/db";
+import { User } from "@/lib/db/schema";
+import { checkAuthorization } from "@/lib/server";
 
 import { NextApiRequest, NextApiResponse } from "next";
 
@@ -10,22 +11,31 @@ export default async function handler(
 ) {
   if (req.method === "POST") {
     try {
+      const isValidAuth = await checkAuthorization(req, res);
+
+      if (!isValidAuth.isValid) {
+        res.status(401).json({
+          message: isValidAuth.message,
+        });
+        return;
+      }
+
       const { email } = req.body;
-      await DBClient.connect();
-      const db = DBClient.db("ITtem");
-      const collection = db.collection("user");
-      const dbUserData = (await collection.findOne({
+
+      await dbConnect();
+
+      const dbUserData = await User.findOne({
         email,
-      })) as UserData | null;
+      });
 
       await deleteToken(dbUserData?.uid || "", "accessToken");
       await deleteToken(dbUserData?.uid || "", "refreshToken");
       res.status(200).json({ message: "성공적으로 토큰이 삭제됬어요." });
     } catch (error) {
       console.error(error);
-      res.status(500).json({ message: "토큰 삭제에 실패했어요." });
-    } finally {
-      await DBClient.close();
+      res.status(500).json({
+        message: "토큰 삭제에 실패했어요.",
+      });
     }
   }
 }
